# api/openapi.yaml
openapi: 3.1.0
info:
  title: Katalog API
  description: >
    REST API сервиса Katalog. Источник правды для контрактов:
    клиентские типы генерируются из этого файла (orval).
    Аутентификация: сессионная httpOnly cookie `katalog_session`.
    Тенант-изоляция: любой чужой или несуществующий ресурс — 404.
  version: 0.5.0
servers:
  - url: /
paths:
...
  /api/v1/photos/{photoID}:
    parameters:
      - name: photoID
        in: path
        required: true
        schema: { type: string, format: uuid }
    patch:
      operationId: updatePhoto
      summary: Обновить подпись фото
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                caption: { type: string, maxLength: 2000 }
      responses:
        "200":
          description: Обновлено
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Photo"
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
    delete:
      operationId: deletePhoto
      summary: Удалить фото (объекты S3, квота, счётчик альбома)
      responses:
        "204": { description: Удалено }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }

  # ---- Публичный неймспейс витрины (без сессии, отдельные схемы ответов:
  # ---- приватные поля наружу не отдаются никогда) ----
...
  /api/v1/public/complaints:
    post:
      operationId: createComplaint
      summary: Жалоба правообладателя на контент (notice-and-takedown), без auth
      description: >
        Жалоба привязывается к магазину/фото, если ссылку удалось распознать
        (/{slug} или /media/{shop_id}/{photo_id}/...); нераспознанный URL
        не мешает подать жалобу. Модератор получает email-уведомление.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [url, reporter_name, reporter_email, reason]
              properties:
                url: { type: string, maxLength: 1000 }
                reporter_name: { type: string, maxLength: 200 }
                reporter_email: { type: string, format: email }
                reason: { type: string, minLength: 10, maxLength: 5000 }
      responses:
        "201":
          description: Жалоба зарегистрирована
          content:
            application/json:
              schema:
                type: object
                required: [id, status]
                properties:
                  id: { type: string, format: uuid }
                  status: { type: string, const: open }
        "400": { $ref: "#/components/responses/BadRequest" }
        "429": { $ref: "#/components/responses/RateLimited" }
  /api/v1/admin/complaints:
    get:
      operationId: adminListComplaints
      summary: "[admin] Список жалоб (фильтр по статусу)"
      description: Все /api/v1/admin/* доступны только role=admin; остальным — 404.
      parameters:
        - name: status
          in: query
          required: false
          schema: { type: string, enum: [open, in_review, resolved, rejected] }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/AdminComplaint" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/v1/admin/complaints/{complaintID}:
    patch:
      operationId: adminUpdateComplaint
      summary: "[admin] Сменить статус жалобы (с записью в аудит-лог)"
      parameters:
        - name: complaintID
          in: path
          required: true
          schema: { type: string, format: uuid }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status: { type: string, enum: [open, in_review, resolved, rejected] }
      responses:
        "200":
          description: Статус обновлён
          content:
            application/json:
              schema:
                type: object
                required: [id, status]
                properties:
                  id: { type: string, format: uuid }
                  status: { type: string }
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
...
components:
...
  responses:
    BadRequest:
      description: Ошибка валидации
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Unauthorized:
      description: Нет или истекла сессия / неверные креды
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    NotFound:
      description: Ресурс не найден или принадлежит другому пользователю
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
    RateLimited:
      description: Превышен rate-limit
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
  schemas:
    Error:
      type: object
      required: [error, message]
      properties:
        error: { type: string, description: Машинный код ошибки }
        message: { type: string }
...
    AdminComplaint:
      type: object
      required: [id, shop_id, shop_slug, photo_id, photo_album_id, reason,
                 reporter_name, reporter_email, content_url, status, created_at, resolved_at]
      properties:
        id: { type: string, format: uuid }
        shop_id: { type: [string, "null"], format: uuid }
        shop_slug: { type: [string, "null"] }
        photo_id: { type: [string, "null"], format: uuid }
        photo_album_id: { type: [string, "null"], format: uuid }
        reason: { type: string }
        reporter_name: { type: string }
        reporter_email: { type: string }
        content_url: { type: string }
        status: { type: string, enum: [open, in_review, resolved, rejected] }
        created_at: { type: string, format: date-time }
        resolved_at: { type: [string, "null"], format: date-time }
...
    Photo:
      type: object
      required: [id, album_id, caption, status, width, height, sort_order]
      properties:
        id: { type: string, format: uuid }
        album_id: { type: string, format: uuid }
        caption: { type: string }
        status: { type: string, enum: [uploading, processing, ready, failed, blocked] }
        width: { type: integer }
        height: { type: integer }
        sort_order: { type: integer }
        urls:
          type: object
          description: Только для status=ready; деривативы WebP (300/800/1600)
          properties:
            thumb: { type: string }
            medium: { type: string }
            large: { type: string }